New File: TCPIPTestAgain
this is modified to do only the connection in parallel.  It has proven
that the socket connections are written to global variables successfully
since we were able to get a message to the server.  However, maybe our
sockets are only one-way? Maybe we should try to create a receive socket
that goes the other way?